Veteran Tweaks ## About The Pull Request
I've noticed we have a lot of veterans picking "Middle-aged" instead of "Old", which means that they're not able to actually TEACH keep roles much at all. This PR aims to fix that.

## Testing Evidence

<!-- It's mandatory to test your PR. Provide images, clips or description of how you tested your changes where possible. -->

## Why It's Good For The Game

Veterans need to do things! Incentivizing them to not do the thing vet is made to do is bad and contributes to so many veterans wasting the slot. EIther we make them old only or let them be middle-aged.
<!-- Argue for the merits of your changes and how they benefit the game, especially if they are controversial. If you can't, then it probably isn't good for the game in the first place. -->